Reorganize github syncing
